refactor(files): rename theme to themeProvider
