+++ lib/clock.js
@@ -1,5 +1,5 @@
 const terminal = require('./terminal')
-const themeProvider = require('./theme.js')
+const themeProvider = require('./themeProvider')
 
 const getPaddingLeft = (theme) => {
   const { width: terminalWidth } = terminal.size()
